+++ lib/service/faceDetectorPainter.dart
@@ -1,0 +1,93 @@
+
+import 'package:flutter/material.dart';
+import 'package:firebase_ml_vision/firebase_ml_vision.dart';
+
+class FaceDetectorPainter extends CustomPainter {
+  FaceDetectorPainter(this.imageSize, this.faces, {this.reflection = false});
+
+  final bool reflection;
+  final Size imageSize;
+  final List<Face> faces;
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final double scaleX = size.width / imageSize.width;
+    final double scaleY = size.height / imageSize.height;
+
+    final Paint paint = Paint()
+      ..style = PaintingStyle.stroke
+      ..strokeWidth = 2.0
+      ..color = Colors.red;
+
+    for (Face face in faces) {
+      final faceRect =
+          _reflectionRect(reflection, face.boundingBox, imageSize.width);
+      if(face.trackingId == 0){
+        paint..color=Colors.green;
+      }
+      else{
+         paint..color=Colors.red;
+      }
+      canvas.drawRect(
+        _scaleRect(
+          rect: faceRect,
+          imageSize: imageSize,
+          widgetSize: size,
+        ),
+        paint,
+      );
+
+      final textStyle = TextStyle(
+        color: Colors.black,
+        fontSize: ((face.boundingBox.right - face.boundingBox.left) * scaleX) *.1,
+        backgroundColor: Color.fromRGBO(255, 255, 255, 0.5)
+      );
+      final textSpan = TextSpan(
+        text: 'Seen ${face.trackingId} Times Today',
+        style: textStyle,
+      );
+      final textPainter = TextPainter(
+        text: textSpan,
+        textDirection: TextDirection.ltr,
+      );
+      textPainter.layout(
+        minWidth: 0,
+        maxWidth: size.width,
+      );
+      final offset = Offset(face.boundingBox.left * scaleX, face.boundingBox.top * scaleY);
+      textPainter.paint(canvas, offset);
+    }
+  }
+
+  @override
+  bool shouldRepaint(FaceDetectorPainter oldDelegate) {
+    return oldDelegate.imageSize != imageSize || oldDelegate.faces != faces;
+  }
+}
+
+Rect _reflectionRect(bool reflection, Rect boundingBox, double width) {
+  if (!reflection) {
+    return boundingBox;
+  }
+  final centerX = width / 2;
+  final left = ((boundingBox.left - centerX) * -1) + centerX;
+  final right = ((boundingBox.right - centerX) * -1) + centerX;
+  return Rect.fromLTRB(left, boundingBox.top, right, boundingBox.bottom);
+}
+
+Rect _scaleRect({
+  @required Rect rect,
+  @required Size imageSize,
+  @required Size widgetSize,
+}) {
+  final scaleX = widgetSize.width / imageSize.width;
+  final scaleY = widgetSize.height / imageSize.height;
+
+  final scaledRect = Rect.fromLTRB(
+    rect.left.toDouble() * scaleX,
+    rect.top.toDouble() * scaleY,
+    rect.right.toDouble() * scaleX,
+    rect.bottom.toDouble() * scaleY,
+  );
+  return scaledRect;
+}
