+++ lib/ui/faceMatchPage.dart
@@ -1,7 +1,11 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:firebase_ml_vision/firebase_ml_vision.dart';
 import 'package:flutter_camera_ml_vision/flutter_camera_ml_vision.dart';
-
+import 'package:tran_dao/service/faceDetectorPainter.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:path/path.dart';
 class FaceMatch extends StatefulWidget {
   @override
   _FaceMatchState createState() => _FaceMatchState();
@@ -16,6 +20,19 @@ class _FaceMatchState extends State<FaceMatch> {
     enableTracking: true,
     mode: FaceDetectorMode.accurate,
   ));
+  Future<String> _requestTempDirectory() async{
+    Directory tempDir = await getTemporaryDirectory();
+    String tempPath = tempDir.path;
+    String filePath = join(tempPath,'Test_capturedFace_${DateTime.now().microsecondsSinceEpoch}.jpg');
+    return filePath;
+  }
+  // Future takePicture(String path,_cameraController) async {
+  //   await _scanKey.currentState._stop(false);
+  //   //FIXME hacky technique to avoid having black screen on some android devices
+  //   await Future.delayed(Duration(milliseconds: 200));
+  //   await _cameraController.takePicture(path);
+  //   _cameraController._start();
+  // }
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -39,8 +56,18 @@ class _FaceMatchState extends State<FaceMatch> {
                   if (faces == null || faces.isEmpty || !mounted) {
                     return;
                   }
+                  _requestTempDirectory().then((path){
+                    print(path);
+                    try {
+                      // _scanKey.currentState.takePicture(path);
+                      // takePicture(path,_scanKey.currentState.cameraController);
+                    } catch (e) {
+                      // takePicture(path,_scanKey.currentState.cameraController);
+                    }
+                  });
                   setState(() {
                     _faces = []..addAll(faces);
+                    print(faces[0].trackingId);
                   });
                 },
                 onDispose: () {
@@ -52,65 +79,4 @@ class _FaceMatchState extends State<FaceMatch> {
       ),
     );
   }
-}
-
-class FaceDetectorPainter extends CustomPainter {
-  FaceDetectorPainter(this.imageSize, this.faces, {this.reflection = false});
-
-  final bool reflection;
-  final Size imageSize;
-  final List<Face> faces;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final Paint paint = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 2.0
-      ..color = Colors.red;
-
-    for (Face face in faces) {
-      final faceRect =
-          _reflectionRect(reflection, face.boundingBox, imageSize.width);
-      canvas.drawRect(
-        _scaleRect(
-          rect: faceRect,
-          imageSize: imageSize,
-          widgetSize: size,
-        ),
-        paint,
-      );
-    }
-  }
-
-  @override
-  bool shouldRepaint(FaceDetectorPainter oldDelegate) {
-    return oldDelegate.imageSize != imageSize || oldDelegate.faces != faces;
-  }
-}
-
-Rect _reflectionRect(bool reflection, Rect boundingBox, double width) {
-  if (!reflection) {
-    return boundingBox;
-  }
-  final centerX = width / 2;
-  final left = ((boundingBox.left - centerX) * -1) + centerX;
-  final right = ((boundingBox.right - centerX) * -1) + centerX;
-  return Rect.fromLTRB(left, boundingBox.top, right, boundingBox.bottom);
-}
-
-Rect _scaleRect({
-  @required Rect rect,
-  @required Size imageSize,
-  @required Size widgetSize,
-}) {
-  final scaleX = widgetSize.width / imageSize.width;
-  final scaleY = widgetSize.height / imageSize.height;
-
-  final scaledRect = Rect.fromLTRB(
-    rect.left.toDouble() * scaleX,
-    rect.top.toDouble() * scaleY,
-    rect.right.toDouble() * scaleX,
-    rect.bottom.toDouble() * scaleY,
-  );
-  return scaledRect;
 }
