+++ pubspec.yaml
@@ -32,8 +32,12 @@ dependencies:
   flutter_launcher_icons: ^0.7.3
   firebase_auth: ^0.15.5+3
   google_sign_in: ^4.4.1 
-  flutter_camera_ml_vision: ^2.2.4
-
+  # flutter_camera_ml_vision: ^2.2.5
+  flutter_camera_ml_vision:
+    git:
+      url: git://github.com/rushio-consulting/flutter_camera_ml_vision.git
+      ref: master
+  
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
